+++ src/js/main.js
@@ -3,7 +3,7 @@
  * see http://siawyoung.com/coding/javascript/exporting-es6-modules-as-single-scripts-with-webpack.html
  */
 import Drawerr from './drawerr';
+import DrawerrMultilevel from './drawerrMultilevel';
 
 window.drawerr = Drawerr;
-
-module.exports = Drawerr;
+window.drawerrMultilevel = DrawerrMultilevel;