+++ changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 2.0.0
+- Add multilevel support with drawerrMultilevel() instance
+- Deprecated init() code moved to constructor
+- Refactoring
+- 
+
 ## 1.1.10
 - Fix: Default options where overwritten instead merged with given args